consolidate tournamentForm and editTournamentForm to one form (same for postForm and editPostForm)
