+++ frontend/src/components/App.js
@@ -10,12 +10,10 @@ import Header from './headerComponents';
 import Footer from './footerComponents';
 import TournamentIndex from './tournamentComponents/TournamentIndex';
 import TournamentShow from './tournamentComponents/tournamentShow';
-import TournamentForm from './formComponents/TournamentForm';
-import EditTournamentForm from './formComponents/EditTournamentForm';
+import TournamentForm from './formComponents/tournamentForm';
 import UserForm from './formComponents/UserForm';
 import Forum from './forumComponents';
 import PostForm from './formComponents/PostForm';
-import EditPostForm from './formComponents/EditPostForm';
 import ProfilePage from './profileComponents';
 
 class App extends Component {
@@ -31,10 +29,10 @@ class App extends Component {
           <Switch>
             <Route exact path="/tournaments/new" component={TournamentForm} />
             <Route exact path="/tournaments/:id" component={TournamentShow} />
-            <Route exact path="/tournaments/:id/edit" component={EditTournamentForm} />
+            <Route exact path="/tournaments/:id/edit" component={TournamentForm} />
             <Route exact path="/tournaments/:id/forum" component={Forum} />
             <Route exact path="/tournaments/:id/forum/new" component={PostForm} />
-            <Route exact path="/tournaments/:id/forum/:postID/edit" component={EditPostForm} />
+            <Route exact path="/tournaments/:id/forum/:postID/edit" component={PostForm} />
             <Route exact path="/tournaments" component={TournamentIndex} />
             <Route exact path="/user/edit" component={UserForm} />
             <Route exact path="/user/profile/:id" component={ProfilePage} />
