+++ frontend/src/components/formComponents/postForm.js
@@ -1,33 +1,54 @@
 import _ from 'lodash';
+import uuidv4 from 'uuid/v4';
 import React, { Component } from 'react';
 import { Field, reduxForm } from 'redux-form';
 import { connect } from 'react-redux';
 import { FIELDS } from '../../util/postFormFields';
 
 //actions
-import { createPost } from '../../actions/postActions';
+import { createPost, editPost } from '../../actions/postActions';
 
 //components
 import TextInput from './formFields/textInput';
-import RequiredTextareaInput from './formFields/requiredTextareaInput';
+import TextareaInput from './formFields/textareaInput';
+import Errors from './formFields/errors';
 import SubmitButton from './formButtons/submitButton';
 import CancelButton from './formButtons/cancelButton';
 
 class PostForm extends Component {
   renderFields() {
-    return _.map(FIELDS, ({ label, name, type }) => {
+    return _.map(FIELDS, ({ label, name, type, placeholder }) => {
       if (name === 'title') {
-        return <Field key={name} component={TextInput} type={type} label={label} name={name} />
+        return <Field key={uuidv4()} component={TextInput}
+          type={type} label={label} name={name}
+        />
       }
 
-      return <Field key={name} component={RequiredTextareaInput} type={type} label={label} name={name} />
+      return (
+        <div key={uuidv4()}>
+          <Field component={TextareaInput}
+            type={type} label={label} name={name} placeholder={placeholder}
+          />
+          <Field component={Errors}
+            type={type} label={label} name={name} placeholder={placeholder}
+          />
+        </div>
+      )
     });
   }
 
   onSubmit(values) {
-    this.props.createPost(this.props.match.params.id, values, () => {
-      this.props.history.push(`/tournaments/${this.props.match.params.id}/forum`);
-    });
+    let splitPath = this.props.match.path.split("/");
+
+    if (splitPath[splitPath.length - 1] === "new") {
+      this.props.createPost(this.props.match.params.id, values, () => {
+        this.props.history.push(`/tournaments/${this.props.match.params.id}/forum`);
+      });
+    } else {
+      this.props.editPost(this.props.postID, values, () => {
+        this.props.history.push(`/tournaments/${this.props.match.params.id}/forum`);
+      });
+    }
   }
 
   render() {
@@ -65,8 +86,12 @@ const validate = values => {
   return errors;
 }
 
+const mapStateToProps = (state, ownProps) => ({
+  postID: ownProps.match.params.postID
+})
+
 export default reduxForm({
   fields: _.keys(FIELDS),
   form: 'PostNewForm',
   validate
-})(connect(null, { createPost })(PostForm));
+})(connect(mapStateToProps, { createPost, editPost })(PostForm));
