+++ frontend/src/components/formComponents/tournamentForm/index.js
@@ -2,22 +2,21 @@ import _ from 'lodash';
 import React, { Component } from 'react';
 import { reduxForm } from 'redux-form';
 import { connect } from 'react-redux'
-import { FIELDS, unrequiredFields } from '../../util/tournamentFormFields';
+import { FIELDS, unrequiredFields } from '../../../util/tournamentFormFields';
 
 //actions
-import { editTournament } from '../../actions/tournamentActions';
+import { createTournament, editTournament } from '../../../actions/tournamentActions';
 
 //components
-import GeneralDetails from './tournamentForm/generalDetails';
-import Location from './tournamentForm/location';
-import Social from './tournamentForm/social';
-import Time from './tournamentForm/time';
-import Fees from './tournamentForm/fees';
-import Description from './tournamentForm/description';
-import SubmitButton from './formButtons/submitButton';
+import GeneralDetails from './generalDetails';
+import Location from './location';
+import Social from './social';
+import Time from './time';
+import Fees from './fees';
+import Description from './description';
+import SubmitButton from '../formButtons/submitButton';
 
-
-class EditTournamentForm extends Component {
+class TournamentForm extends Component {
   onCancel() {
     if (window.confirm("Leaving this page before submitting will result in losing all progress made on this form. Are you sure you would like to cancel? ")) {
       this.props.history.push('/tournaments');
@@ -25,14 +24,21 @@ class EditTournamentForm extends Component {
   }
 
   onSubmit(values) {
-    this.props.editTournament(this.props.match.params.id, values, () => {
-      this.props.history.push(`/tournaments/${this.props.match.params.id}`);
-    });
+    let splitPath = this.props.match.path.split("/");
+
+    if (splitPath[splitPath.length - 1] === "new") {
+      this.props.createTournament(values, () => {
+        this.props.history.push('/tournaments');
+      });
+    } else {
+      this.props.editTournament(this.props.match.params.id, values, () => {
+        this.props.history.push(`/tournaments/${this.props.match.params.id}`);
+      });
+    }
   }
 
   render() {
     const { handleSubmit } = this.props;
-
     return (
       <div className="tournament-form-container">
         <h2>Tournament Form</h2>
@@ -48,7 +54,9 @@ class EditTournamentForm extends Component {
           <Description />
           <div className="form-buttons">
             <SubmitButton />
-            <button className="cancel-button" onClick={this.onCancel.bind(this)}>Cancel</button>
+            <button className="cancel-button"
+              onClick={this.onCancel.bind(this)}
+            >Cancel</button>
           </div>
         </form>
       </div>
@@ -68,6 +76,10 @@ const validate = values => {
     }
   });
 
+  if (_.keys(values).indexOf("seriesDay") === -1) {
+    errors["seriesDay"] = `Please enter the day of the week your tournament will be held on.`;
+  }
+
   return errors;
 }
 
@@ -75,4 +87,4 @@ export default reduxForm({
   fields: _.keys(FIELDS),
   form: 'TournamentNewForm',
   validate
-})(connect(null, { editTournament })(EditTournamentForm));
+})(connect(null, { createTournament, editTournament })(TournamentForm));
